\name{pathwy.A.up}
\alias{pathwy.A.up}
\docType{data}
\title{
Pathway or function data
}
\description{
pathway data are functions or pathways of up-regulated 
genes given by gene annotation tool such as DAVID 
gene function annotion.
}
\usage{data("pathwy.A.up")}
\format{
A data frame with 39 observations on the following 159 
variables.
  \describe{
     \item{\code{V1}}{a factor with levels}
    \item{\code{V2}}{a factor with levels}
    \item{\code{V3}}{a factor with levels}
    \item{\code{V4}}{a factor with levels}
    \item{\code{V5}}{a factor with levels}
    \item{\code{V6}}{a factor with levels}
    \item{\code{V7}}{a factor with levels}
    \item{\code{V8}}{a factor with levels} 
    \item{\code{V9}}{a factor with levels}
    \item{\code{V10}}{a factor with levels}
    \item{\code{V11}}{a factor with levels}
    \item{\code{V12}}{a factor with levels}
    \item{\code{V13}}{a factor with levels}
    \item{\code{V14}}{a factor with levels}
    \item{\code{V15}}{a factor with levels}
    \item{\code{V16}}{a factor with levels} 
    \item{\code{V17}}{a factor with levels}
    \item{\code{V18}}{a factor with levels}
    \item{\code{V19}}{a factor with levels}
    \item{\code{V20}}{a factor with levels}
    \item{\code{V21}}{a factor with levels}
    \item{\code{V22}}{a factor with levels}
    \item{\code{V23}}{a factor with levels}
    \item{\code{V24}}{a factor with levels} 
    \item{\code{V25}}{a factor with levels}
    \item{\code{V26}}{a factor with levels}
    \item{\code{V27}}{a factor with levels}
    \item{\code{V28}}{a factor with levels}
    \item{\code{V29}}{a factor with levels}
    \item{\code{V30}}{a factor with levels}
    \item{\code{V31}}{a factor with levels}
    \item{\code{V32}}{a factor with levels}     
    \item{\code{V33}}{a factor with levels}
    \item{\code{V34}}{a factor with levels}
    \item{\code{V35}}{a factor with levels}
    \item{\code{V36}}{a factor with levels}
    \item{\code{V37}}{a factor with levels}
    \item{\code{V38}}{a factor with levels}
    \item{\code{V39}}{a factor with levels}
    \item{\code{V40}}{a factor with levels} 
    \item{\code{V41}}{a factor with levels}
    \item{\code{V42}}{a factor with levels}
    \item{\code{V43}}{a factor with levels}
    \item{\code{V44}}{a factor with levels}
    \item{\code{V45}}{a factor with levels}
    \item{\code{V46}}{a factor with levels}
    \item{\code{V47}}{a factor with levels}
    \item{\code{V48}}{a factor with levels} 
    \item{\code{V49}}{a factor with levels}
    \item{\code{V50}}{a factor with levels}
    \item{\code{V51}}{a factor with levels}
    \item{\code{V52}}{a factor with levels}
    \item{\code{V53}}{a factor with levels}
    \item{\code{V54}}{a factor with levels}
    \item{\code{V55}}{a factor with levels}
    \item{\code{V56}}{a factor with levels} 
    \item{\code{V57}}{a factor with levels}
    \item{\code{V58}}{a factor with levels}
    \item{\code{V59}}{a factor with levels}
    \item{\code{V60}}{a factor with levels}
    \item{\code{V61}}{a factor with levels}
    \item{\code{V62}}{a factor with levels}
    \item{\code{V63}}{a factor with levels}     
    \item{\code{V64}}{a factor with levels}
    \item{\code{V65}}{a factor with levels}
    \item{\code{V66}}{a factor with levels}
    \item{\code{V67}}{a factor with levels}
    \item{\code{V68}}{a factor with levels}
    \item{\code{V69}}{a factor with levels}
    \item{\code{V70}}{a factor with levels}
    \item{\code{V71}}{a factor with levels} 
    \item{\code{V72}}{a factor with levels}
    \item{\code{V73}}{a factor with levels}
    \item{\code{V74}}{a factor with levels}
    \item{\code{V75}}{a factor with levels}
    \item{\code{V76}}{a factor with levels}
    \item{\code{V77}}{a factor with levels}
    \item{\code{V78}}{a factor with levels}
    \item{\code{V79}}{a factor with levels} 
    \item{\code{V80}}{a factor with levels}
    \item{\code{V81}}{a factor with levels}
    \item{\code{V82}}{a factor with levels}
    \item{\code{V83}}{a factor with levels}
    \item{\code{V84}}{a factor with levels}
    \item{\code{V85}}{a factor with levels}
    \item{\code{V86}}{a factor with levels}
    \item{\code{V87}}{a factor with levels} 
    \item{\code{V88}}{a factor with levels}
    \item{\code{V89}}{a factor with levels}
    \item{\code{V90}}{a factor with levels}
    \item{\code{V91}}{a factor with levels}
    \item{\code{V92}}{a factor with levels}
    \item{\code{V93}}{a factor with levels}
    \item{\code{V94}}{a factor with levels}
    \item{\code{V95}}{a factor with levels}     
    \item{\code{V96}}{a factor with levels}
    \item{\code{V97}}{a factor with levels}
    \item{\code{V98}}{a factor with levels}
    \item{\code{V99}}{a factor with levels}
    \item{\code{V100}}{a factor with levels}
    \item{\code{V101}}{a factor with levels}
    \item{\code{V102}}{a factor with levels}
    \item{\code{V103}}{a factor with levels} 
    \item{\code{V104}}{a factor with levels}
    \item{\code{V105}}{a factor with levels}
    \item{\code{V106}}{a factor with levels}
    \item{\code{V107}}{a factor with levels}
    \item{\code{V108}}{a factor with levels}
    \item{\code{V109}}{a factor with levels}
    \item{\code{V110}}{a factor with levels}
    \item{\code{V111}}{a factor with levels} 
    \item{\code{V112}}{a factor with levels}
    \item{\code{V113}}{a factor with levels}
    \item{\code{V114}}{a factor with levels}
    \item{\code{V115}}{a factor with levels}
    \item{\code{V116}}{a factor with levels}
    \item{\code{V117}}{a factor with levels}
    \item{\code{V118}}{a factor with levels}
    \item{\code{V119}}{a factor with levels} 
    \item{\code{V120}}{a factor with levels}
    \item{\code{V121}}{a factor with levels}
    \item{\code{V122}}{a factor with levels}
    \item{\code{V123}}{a factor with levels}
    \item{\code{V124}}{a factor with levels}
    \item{\code{V125}}{a factor with levels}
    \item{\code{V126}}{a factor with levels}
    \item{\code{V127}}{a factor with levels}        
    \item{\code{V128}}{a factor with levels}
    \item{\code{V129}}{a factor with levels} 
    \item{\code{V130}}{a factor with levels}
    \item{\code{V131}}{a factor with levels}
    \item{\code{V132}}{a factor with levels}
    \item{\code{V133}}{a factor with levels}
    \item{\code{V134}}{a factor with levels}
    \item{\code{V135}}{a factor with levels}
    \item{\code{V136}}{a factor with levels}
    \item{\code{V137}}{a factor with levels} 
    \item{\code{V138}}{a factor with levels}
    \item{\code{V139}}{a factor with levels}
    \item{\code{V140}}{a factor with levels}
    \item{\code{V141}}{a factor with levels}
    \item{\code{V142}}{a factor with levels}
    \item{\code{V143}}{a factor with levels}
    \item{\code{V144}}{a factor with levels}
    \item{\code{V145}}{a factor with levels} 
    \item{\code{V146}}{a factor with levels}
    \item{\code{V147}}{a factor with levels}
    \item{\code{V148}}{a factor with levels}
    \item{\code{V149}}{a factor with levels}
    \item{\code{V150}}{a factor with levels}
    \item{\code{V151}}{a factor with levels}
    \item{\code{V152}}{a factor with levels}
    \item{\code{V153}}{a factor with levels}
    \item{\code{V154}}{a factor with levels}
    \item{\code{V155}}{a factor with levels}
    \item{\code{V156}}{a factor with levels}
    \item{\code{V157}}{a factor with levels}
    \item{\code{V158}}{a factor with levels}
    \item{\code{V159}}{a factor with levels}  	
              
}
}
\details{
The 39 observations are 39 functions/pathways 
listed each being a set up-regulated genes found 
by NBBttest in knockdowned DDX39 cell line. Log10 
of pvalue of gene in differential expression 
detection is used to weight conpoment of this 
gene playing role in this function. Then heatmap 
of pathways or functions are made using myheatmap. 
V1-V159 are gene sets of function or pathways   
}
\examples{
data(pathwy.A.up)
}
\keyword{datasets}
