\name{skjt}
\alias{skjt}
\docType{data}
\title{
Simulated Null Transcriptomic data
}
\description{
The dataset skjt generated by using R negative binomial 
pseudorandom generator rnbinom is used as an example 
for calculating omega.
}
\usage{data("skjt")}
\format{
A data frame with 13409 observations on the following 
14 variables.
  \describe{
    \item{\code{geneid}}{: a string vector }
    \item{\code{tagid}}{: a numeric vector}
    \item{\code{geneid.1}}{: a numeric vector}
    \item{\code{name}}{: a string vector}
    \item{\code{chr}}{: a string vector}
    \item{\code{strand}}{: a character vector}
    \item{\code{pos}}{: a numeric vector}
    \item{\code{anno}}{: a string vector }
    \item{\code{Jurk.NS.A}}{: a numeric vector}
    \item{\code{Jurk.NS.B}}{: a numeric vector}
    \item{\code{Jurk.NS.C}}{: a numeric vector}
    \item{\code{Jurk.48h.A}}{: a numeric vector}
    \item{\code{Jurk.48h.B}}{: a numeric vector}
    \item{\code{Jurk.48h.C}}{: a numeric vector}
  }
}
\details{
The dataset skjt was generated by using R negative 
binomial pseudorandom generator rnbinom with \eqn{\mu} 
and size. Parameters \eqn{\mu} and size are given 
by mean and  variance drawn from real  Jurkat T cell 
transcriptomic count data . Condition (or treatment) 
effect on differential transcription of isoforms was 
set to zero. The data have 13409 genes and 7 information 
columns: geneid	 tagid	name \code{chr},\code{strand},
\code{pos},\code{anno}, and 6 data columns.
}

\value{
a datasheet containd ID, information, count data of 
RNA reads.
	}
\references{
Yuan-De Tan Anita M. Chandler, Arindam Chaudhury, and 
Joel R. Neilson(2015) A Powerful Statistical Approach 
for Large-scale Differential Transcription Analysis. 
\emph{Plos One}. DOI: 10.1371/journal.pone.0123658. 
}
\examples{
data(skjt)
}
\keyword{datasets}
